Mod: component 추가구성
