--- src/pages/41_dahye/Main/Comment.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Comment = ({ i }) => {
-  return <li>{i}</li>;
-};
-
-export default Comment;
