+++ src/pages/41_dahye/Main/Main.scss
@@ -250,7 +250,7 @@
 
       .story-box {
         margin: 25px 0px 0px 0px;
-        padding: 0px 10px 0px 10px;
+        padding: 0px 10px 10px 10px;
         border: 1px solid gray;
       }
 
@@ -289,6 +289,12 @@
           margin: 0;
           margin-top: 7px;
         }
+      }
+
+      .aside {
+        margin: 25px 0px 0px 0px;
+        padding: 0px 10px 0px 10px;
+        border: 1px solid gray;
       }
     }
   }
