+++ src/pages/41_dahye/Main/components/Aside.js
@@ -1,0 +1,8 @@
+import React, { useState } from 'react';
+import '../Main.scss';
+
+const Aside = () => {
+  return <div className="aside">box</div>;
+};
+
+export default Aside;
