+++ src/pages/41_dahye/Main/components/Comment.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import '../Main.scss';
+
+const Comment = ({ item }) => {
+  return <li>{item}</li>;
+};
+
+export default Comment;
