+++ src/pages/41_dahye/Main/components/Feeds.js
@@ -1,0 +1,16 @@
+import React, { useState } from 'react';
+// import Comment from '../components/Comment.js';
+import Feed from './Feed.js';
+import Mainright from './Mainright.js';
+import '../Main.scss';
+
+const Feeds = () => {
+  return (
+    <main className="main">
+      <Feed />
+      <Mainright />
+    </main>
+  );
+};
+
+export default Feeds;
