+++ src/pages/41_dahye/Main/components/Nav.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Main.scss';
+import '../Main.scss';
 
 const Nav = () => {
   return (
